Refactor Layout.tsx
